Fix ai helpers and put in own file
